+++ newsletter/admin.py
@@ -1,14 +1,12 @@
-from __future__ import unicode_literals
-
 import logging
-logger = logging.getLogger(__name__)
 
-import six
+from django.urls import path
+
+logger = logging.getLogger(__name__)
 
 from django.db import models
 
 from django.conf import settings
-from django.conf.urls import url
 
 from django.contrib import admin, messages
 from django.contrib.sites.models import Site
@@ -21,7 +19,7 @@ from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render
 
 from django.utils.html import format_html
-from django.utils.translation import ugettext as _, ungettext
+from django.utils.translation import gettext as _, ngettext
 from django.utils.formats import date_format
 
 from django.views.decorators.clickjacking import xframe_options_sameorigin
@@ -32,28 +30,36 @@ except ImportError:  # Django < 1.10
     from django.views.i18n import javascript_catalog
     HAS_CBV_JSCAT = False
 
-from sorl.thumbnail.admin import AdminImageMixin
+# Conditional imports as only one Thumbnail app is required
+try:
+    from sorl.thumbnail.admin import AdminImageMixin
+except ImportError:
+    pass
+
+try:
+    from easy_thumbnails.widgets import ImageClearableFileInput
+except (ImportError, RuntimeError):
+    pass
 
 from .models import (
-    Newsletter, Subscription, Article, Message, Submission
+    Newsletter, Subscription, Attachment, Article, Message, Submission
 )
 
 from django.utils.timezone import now
+from django.urls import reverse
 
 from .admin_forms import (
     SubmissionAdminForm, SubscriptionAdminForm, ImportForm, ConfirmForm,
     ArticleFormSet
 )
 from .admin_utils import ExtendibleModelAdminMixin, make_subscription
-
-from .compat import get_context, reverse
-
+from .fields import DynamicImageField
 from .settings import newsletter_settings
 
 from django.forms import HiddenInput  # custom
 from djangocms_text_ckeditor.widgets import TextEditorWidget  # custom
 
-# Contsruct URL's for icons
+# Construct URL's for icons
 ICON_URLS = {
     'yes': '%snewsletter/admin/img/icon-yes.gif' % settings.STATIC_URL,
     'wait': '%snewsletter/admin/img/waiting.gif' % settings.STATIC_URL,
@@ -93,7 +99,7 @@ class NewsletterAdmin(admin.ModelAdmin):
     admin_submissions.short_description = ''
 
 
-class NewsletterAdminLinkMixin(object):
+class NewsletterAdminLinkMixin:
     def admin_newsletter(self, obj):
         opts = Newsletter._meta
         newsletter = obj.newsletter
@@ -187,17 +193,26 @@ class SubmissionAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
 
     """ URLs """
     def get_urls(self):
-        urls = super(SubmissionAdmin, self).get_urls()
+        urls = super().get_urls()
 
         my_urls = [
-            url(
-                r'^(.+)/submit/$',
+            path(
+                '<object_id>/submit/',
                 self._wrap(self.submit),
                 name=self._view_name('submit')
             )
         ]
 
         return my_urls + urls
+
+
+class AttachmentInline(admin.TabularInline):
+    model = Attachment
+    extra = 1
+
+    def has_change_permission(self, request, obj=None):
+        """ Prevent change of the file (instead needs to be deleted) """
+        return False
 
 
 StackedInline = admin.StackedInline
@@ -218,8 +233,17 @@ if (
             'Imperavi WYSIWYG text editor might not work.'
         )
 
+# Creates a base class for the ArticleInline to inherit depending on
+# if the user has decided to use sorl-thumbnail or not
+# https://sorl-thumbnail.readthedocs.io/en/latest/examples.html#admin-examples
+if newsletter_settings.THUMBNAIL == 'sorl-thumbnail':
+    ArticleInlineClassTuple = (AdminImageMixin, StackedInline)
+else:
+    ArticleInlineClassTuple = (StackedInline,)
 
-class ArticleInline(AdminImageMixin, StackedInline):
+BaseArticleInline = type('BaseArticleInline', ArticleInlineClassTuple, {})
+
+class ArticleInline(BaseArticleInline):
     model = Article
     extra = 0
     formset = ArticleFormSet
@@ -233,10 +257,19 @@ class ArticleInline(AdminImageMixin, StackedInline):
         }),
     )
 
+    # Perform any formfield overrides depending on specified settings
+    formfield_overrides = {}
+
     if newsletter_settings.RICHTEXT_WIDGET:
         formfield_overrides = {
             models.TextField: {'widget': TextEditorWidget(
                 configuration='CKEDITOR_SETTINGS_BASE')},
+        }
+
+    # https://easy-thumbnails.readthedocs.io/en/latest/usage/#forms
+    if newsletter_settings.THUMBNAIL == 'easy-thumbnails':
+        formfield_overrides[DynamicImageField] = {
+            'widget': ImageClearableFileInput
         }
 
 
@@ -251,7 +284,7 @@ class MessageAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
     date_hierarchy = 'date_create'
     prepopulated_fields = {'slug': ('title',)}
 
-    inlines = [ArticleInline, ]
+    inlines = [ArticleInline, AttachmentInline, ]
 
     """ List extensions """
     def admin_title(self, obj):
@@ -269,7 +302,8 @@ class MessageAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
         return render(
             request,
             "admin/newsletter/message/preview.html",
-            {'message': self._getobj(request, object_id)},
+            {'message': self._getobj(request, object_id),
+             'attachments': Attachment.objects.filter(message_id=object_id)},
         )
 
     @xframe_options_sameorigin
@@ -282,12 +316,14 @@ class MessageAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
                 'message belongs to.'
             ))
 
-        c = get_context({'message': message,
-                         'site': Site.objects.get_current(),
-                         'newsletter': message.newsletter,
-                         'date': now(),
-                         'STATIC_URL': settings.STATIC_URL,
-                         'MEDIA_URL': settings.MEDIA_URL})
+        c = {
+            'message': message,
+            'site': Site.objects.get_current(),
+            'newsletter': message.newsletter,
+            'date': now(),
+            'STATIC_URL': settings.STATIC_URL,
+            'MEDIA_URL': settings.MEDIA_URL
+        }
 
         return HttpResponse(message.html_template.render(c))
 
@@ -295,14 +331,14 @@ class MessageAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
     def preview_text(self, request, object_id):
         message = self._getobj(request, object_id)
 
-        c = get_context({
+        c = {
             'message': message,
             'site': Site.objects.get_current(),
             'newsletter': message.newsletter,
             'date': now(),
             'STATIC_URL': settings.STATIC_URL,
             'MEDIA_URL': settings.MEDIA_URL
-        }, autoescape=False)
+        }
 
         return HttpResponse(
             message.text_template.render(c),
@@ -327,24 +363,24 @@ class MessageAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
 
     """ URLs """
     def get_urls(self):
-        urls = super(MessageAdmin, self).get_urls()
+        urls = super().get_urls()
 
         my_urls = [
-            url(r'^(.+)/preview/$',
-                self._wrap(self.preview),
-                name=self._view_name('preview')),
-            url(r'^(.+)/preview/html/$',
-                self._wrap(self.preview_html),
-                name=self._view_name('preview_html')),
-            url(r'^(.+)/preview/text/$',
-                self._wrap(self.preview_text),
-                name=self._view_name('preview_text')),
-            url(r'^(.+)/submit/$',
-                self._wrap(self.submit),
-                name=self._view_name('submit')),
-            url(r'^(.+)/subscribers/json/$',
-                self._wrap(self.subscribers_json),
-                name=self._view_name('subscribers_json')),
+            path('<object_id>/preview/',
+                 self._wrap(self.preview),
+                 name=self._view_name('preview')),
+            path('<object_id>/preview/html/',
+                 self._wrap(self.preview_html),
+                 name=self._view_name('preview_html')),
+            path('<object_id>/preview/text/',
+                 self._wrap(self.preview_text),
+                 name=self._view_name('preview_text')),
+            path('<object_id>/submit/',
+                 self._wrap(self.submit),
+                 name=self._view_name('submit')),
+            path('<object_id>/subscribers/json/',
+                 self._wrap(self.subscribers_json),
+                 name=self._view_name('subscribers_json')),
         ]
 
         return my_urls + urls
@@ -413,7 +449,7 @@ class SubscriptionAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
         rows_updated = queryset.update(subscribed=True)
         self.message_user(
             request,
-            ungettext(
+            ngettext(
                 "%d user has been successfully subscribed.",
                 "%d users have been successfully subscribed.",
                 rows_updated
@@ -425,7 +461,7 @@ class SubscriptionAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
         rows_updated = queryset.update(subscribed=False)
         self.message_user(
             request,
-            ungettext(
+            ngettext(
                 "%d user has been successfully unsubscribed.",
                 "%d users have been successfully unsubscribed.",
                 rows_updated
@@ -475,7 +511,7 @@ class SubscriptionAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
             form = ConfirmForm(request.POST)
             if form.is_valid():
                 try:
-                    for email, name in six.iteritems(addresses):
+                    for email, name in addresses.items():
                         address_inst = make_subscription(
                             newsletter, email, name
                         )
@@ -486,7 +522,7 @@ class SubscriptionAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
 
                 messages.success(
                     request,
-                    ungettext(
+                    ngettext(
                         "%d subscription has been successfully added.",
                         "%d subscriptions have been successfully added.",
                         len(addresses)
@@ -508,27 +544,27 @@ class SubscriptionAdmin(NewsletterAdminLinkMixin, ExtendibleModelAdminMixin,
 
     """ URLs """
     def get_urls(self):
-        urls = super(SubscriptionAdmin, self).get_urls()
+        urls = super().get_urls()
 
         my_urls = [
-            url(r'^import/$',
-                self._wrap(self.subscribers_import),
-                name=self._view_name('import')),
-            url(r'^import/confirm/$',
-                self._wrap(self.subscribers_import_confirm),
-                name=self._view_name('import_confirm')),
+            path('import/',
+                 self._wrap(self.subscribers_import),
+                 name=self._view_name('import')),
+            path('import/confirm/',
+                 self._wrap(self.subscribers_import_confirm),
+                 name=self._view_name('import_confirm')),
         ]
         # Translated JS strings - these should be app-wide but are
         # only used in this part of the admin. For now, leave them here.
         if HAS_CBV_JSCAT:
-            my_urls.append(url(r'^jsi18n/$',
-                JavaScriptCatalog.as_view(packages=('newsletter',)),
-                name='newsletter_js18n'))
+            my_urls.append(path('jsi18n/',
+                           JavaScriptCatalog.as_view(packages=('newsletter',)),
+                           name='newsletter_js18n'))
         else:
-            my_urls.append(url(r'^jsi18n/$',
-                javascript_catalog,
-                {'packages': ('newsletter',)},
-                name='newsletter_js18n'))
+            my_urls.append(path('jsi18n/',
+                                javascript_catalog,
+                                {'packages': ('newsletter',)},
+                                name='newsletter_js18n'))
 
         return my_urls + urls