+++ newsletter/models.py
@@ -1,6 +1,8 @@
 import logging
+import os
 import time
-from six import python_2_unicode_compatible
+from datetime import datetime
+
 import django
 
 from django.conf import settings
@@ -10,15 +12,14 @@ from django.core.mail import EmailMultiAlternatives
 from django.db import models
 from django.template.loader import select_template
 from django.utils.functional import cached_property
-from django.utils.translation import ugettext_lazy as _
-from django.utils.translation import ugettext
+from django.utils.translation import gettext_lazy as _
+from django.utils.translation import gettext
 from django.utils.timezone import now
+from django.urls import reverse
 
-from sorl.thumbnail import ImageField
 from distutils.version import LooseVersion
 
-
-from .compat import get_context, reverse
+from .fields import DynamicImageField
 from .utils import (
     make_activation_code, get_default_sites, ACTIONS
 )
@@ -31,7 +32,6 @@ logger = logging.getLogger(__name__)
 AUTH_USER_MODEL = getattr(settings, 'AUTH_USER_MODEL', 'auth.User')
 
 
-@python_2_unicode_compatible
 class Newsletter(models.Model):
     site = models.ManyToManyField(Site, default=get_default_sites)
 
@@ -98,7 +98,7 @@ class Newsletter(models.Model):
             # HTML templates are not required
             html_template = None
 
-        return (subject_template, text_template, html_template)
+        return subject_template, text_template, html_template
 
     def __str__(self):
         return self.title
@@ -126,7 +126,7 @@ class Newsletter(models.Model):
         return get_address(self.sender, self.email)
 
     def get_subscriptions(self):
-        logger.debug(u'Looking up subscribers for %s', self)
+        logger.debug('Looking up subscribers for %s', self)
 
         return Subscription.objects.filter(newsletter=self, subscribed=True)
 
@@ -138,7 +138,6 @@ class Newsletter(models.Model):
             return None
 
 
-@python_2_unicode_compatible
 class Subscription(models.Model):
     user = models.ForeignKey(
         AUTH_USER_MODEL, blank=True, null=True, verbose_name=_('user'),
@@ -146,7 +145,7 @@ class Subscription(models.Model):
     )
 
     name_field = models.CharField(
-        db_column='name', max_length=30, blank=True, null=True,
+        db_column='name', max_length=200, blank=True, null=True,
         verbose_name=_('name'), help_text=_('optional')
     )
 
@@ -191,7 +190,7 @@ class Subscription(models.Model):
             self.unsubscribed = True
 
         logger.debug(
-            _(u'Updated subscription %(subscription)s to %(action)s.'),
+            _('Updated subscription %(subscription)s to %(action)s.'),
             {
                 'subscription': self,
                 'action': action
@@ -207,7 +206,7 @@ class Subscription(models.Model):
         Internal helper method for managing subscription state
         during subscription.
         """
-        logger.debug(u'Subscribing subscription %s.', self)
+        logger.debug('Subscribing subscription %s.', self)
 
         self.subscribe_date = now()
         self.subscribed = True
@@ -218,7 +217,7 @@ class Subscription(models.Model):
         Internal helper method for managing subscription state
         during unsubscription.
         """
-        logger.debug(u'Unsubscribing subscription %s.', self)
+        logger.debug('Unsubscribing subscription %s.', self)
 
         self.subscribed = False
         self.unsubscribed = True
@@ -276,7 +275,7 @@ class Subscription(models.Model):
             elif self.unsubscribed:
                 self._unsubscribe()
 
-        super(Subscription, self).save(*args, **kwargs)
+        super().save(*args, **kwargs)
 
     ip = models.GenericIPAddressField(_("IP address"), blank=True, null=True)
 
@@ -308,14 +307,14 @@ class Subscription(models.Model):
 
     def __str__(self):
         if self.name:
-            return _(u"%(name)s <%(email)s> to %(newsletter)s") % {
+            return _("%(name)s <%(email)s> to %(newsletter)s") % {
                 'name': self.name,
                 'email': self.email,
                 'newsletter': self.newsletter
             }
 
         else:
-            return _(u"%(email)s to %(newsletter)s") % {
+            return _("%(email)s to %(newsletter)s") % {
                 'email': self.email,
                 'newsletter': self.newsletter
             }
@@ -343,10 +342,8 @@ class Subscription(models.Model):
             'MEDIA_URL': settings.MEDIA_URL
         }
 
-        unescaped_context = get_context(variable_dict, autoescape=False)
-
-        subject = subject_template.render(unescaped_context).strip()
-        text = text_template.render(unescaped_context)
+        subject = subject_template.render(variable_dict).strip()
+        text = text_template.render(variable_dict)
 
         message = EmailMultiAlternatives(
             subject, text,
@@ -355,17 +352,15 @@ class Subscription(models.Model):
         )
 
         if html_template:
-            escaped_context = get_context(variable_dict)
-
             message.attach_alternative(
-                html_template.render(escaped_context), "text/html"
+                html_template.render(variable_dict), "text/html"
             )
 
         message.send()
 
         logger.debug(
-            u'Activation email sent for action "%(action)s" to %(subscriber)s '
-            u'with activation code "%(action_code)s".', {
+            'Activation email sent for action "%(action)s" to %(subscriber)s '
+            'with activation code "%(action_code)s".', {
                 'action_code': self.activation_code,
                 'action': action,
                 'subscriber': self
@@ -397,7 +392,6 @@ class Subscription(models.Model):
         })
 
 
-@python_2_unicode_compatible
 class Article(models.Model):
     """
     An Article within a Message which will be send through a Submission.
@@ -422,7 +416,7 @@ class Article(models.Model):
     )
 
     # Make this a foreign key for added elegance
-    image = ImageField(
+    image = DynamicImageField(
         upload_to='newsletter/images/%Y/%m/%d', blank=True, null=True,
         verbose_name=_('image')
     )
@@ -452,13 +446,49 @@ class Article(models.Model):
             # as to assure uniqueness.
             self.sortorder = self.post.get_next_article_sortorder()
 
-        super(Article, self).save(*args, **kwargs)
+        super().save()
+
+
+def attachment_upload_to(instance, filename):
+    return os.path.join(
+        'newsletter', 'attachments',
+        datetime.utcnow().strftime('%Y-%m-%d'),
+        str(instance.message.id),
+        filename
+    )
+
+
+class Attachment(models.Model):
+    """ Attachment for a Message. """
+    class Meta:
+        verbose_name = _('attachment')
+        verbose_name_plural = _('attachments')
+
+    def __str__(self):
+        return _("%(file_name)s on %(message)s") % {
+            'file_name': self.file_name,
+            'message': self.message
+        }
+
+    file = models.FileField(
+        upload_to=attachment_upload_to,
+        blank=False, null=False,
+        verbose_name=_('attachment')
+    )
+
+    message = models.ForeignKey(
+        'Message', verbose_name=_('message'), on_delete=models.CASCADE, related_name='attachments',
+    )
+
+    @property
+    def file_name(self):
+        return os.path.split(self.file.name)[1]
 
 
 def get_default_newsletter():
     return Newsletter.get_default()
 
-@python_2_unicode_compatible
+
 class Message(models.Model):
     """ Message as sent through a Submission. """
 
@@ -486,7 +516,7 @@ class Message(models.Model):
 
     def __str__(self):
         try:
-            return _(u"%(title)s in %(newsletter)s") % {
+            return _("%(title)s in %(newsletter)s") % {
                 'title': self.title,
                 'newsletter': self.newsletter
             }
@@ -531,7 +561,6 @@ class Message(models.Model):
             return None
 
 
-@python_2_unicode_compatible
 class Submission(models.Model):
     """
     Submission represents a particular Message as it is being submitted
@@ -543,7 +572,7 @@ class Submission(models.Model):
         verbose_name_plural = _('submissions')
 
     def __str__(self):
-        return _(u"%(newsletter)s on %(publish_date)s") % {
+        return _("%(newsletter)s on %(publish_date)s") % {
             'newsletter': self.message,
             'publish_date': self.publish_date
         }
@@ -562,7 +591,7 @@ class Submission(models.Model):
         subscriptions = self.subscriptions.filter(subscribed=True)
 
         logger.info(
-            ugettext(u"Submitting %(submission)s to %(count)d people"),
+            gettext("Submitting %(submission)s to %(count)d people"),
             {'submission': self, 'count': subscriptions.count()}
         )
 
@@ -598,11 +627,9 @@ class Submission(models.Model):
             'MEDIA_URL': settings.MEDIA_URL
         }
 
-        unescaped_context = get_context(variable_dict, autoescape=False)
-
         subject = self.message.subject_template.render(
-            unescaped_context).strip()
-        text = self.message.text_template.render(unescaped_context)
+            variable_dict).strip()
+        text = self.message.text_template.render(variable_dict)
 
         message = EmailMultiAlternatives(
             subject, text,
@@ -611,17 +638,20 @@ class Submission(models.Model):
             headers=self.extra_headers,
         )
 
-        if self.message.html_template:
-            escaped_context = get_context(variable_dict)
+        attachments = Attachment.objects.filter(message_id=self.message.id)
 
+        for attachment in attachments:
+            message.attach_file(attachment.file.path)
+
+        if self.message.html_template:
             message.attach_alternative(
-                self.message.html_template.render(escaped_context),
+                self.message.html_template.render(variable_dict),
                 "text/html"
             )
 
         try:
             logger.debug(
-                ugettext(u'Submitting message to: %s.'),
+                gettext('Submitting message to: %s.'),
                 subscription
             )
 
@@ -630,8 +660,8 @@ class Submission(models.Model):
         except Exception as e:
             # TODO: Test coverage for this branch.
             logger.error(
-                ugettext(u'Message %(subscription)s failed '
-                         u'with error: %(error)s'),
+                gettext('Message %(subscription)s failed '
+                        'with error: %(error)s'),
                 {'subscription': subscription,
                  'error': e}
             )
@@ -648,7 +678,7 @@ class Submission(models.Model):
 
     @classmethod
     def from_message(cls, message):
-        logger.debug(ugettext('Submission of message %s'), message)
+        logger.debug(gettext('Submission of message %s'), message)
         submission = cls()
         submission.message = message
         submission.newsletter = message.newsletter
@@ -665,9 +695,7 @@ class Submission(models.Model):
 
         self.newsletter = self.message.newsletter
 
-        return super(Submission, self).save()
-
-
+        return super().save()
 
     def get_absolute_url(self):
         assert self.newsletter.slug
@@ -722,12 +750,13 @@ class Submission(models.Model):
         db_index=True, editable=False
     )
 
+
 def get_address(name, email):
     # Converting name to ascii for compatibility with django < 1.9.
     # Remove this when django 1.8 is no longer supported.
     if LooseVersion(django.get_version()) < LooseVersion('1.9'):
         name = name.encode('ascii', 'ignore').decode('ascii').strip()
     if name:
-        return u'%s <%s>' % (name, email)
+        return '%s <%s>' % (name, email)
     else:
-        return u'%s' % email
+        return '%s' % email