+++ newsletter/views.py
@@ -22,12 +22,12 @@ from django.contrib.sites.models import Site
 from django.contrib.auth.decorators import login_required
 
 from django.utils.decorators import method_decorator
-from django.utils.translation import ugettext, ugettext_lazy as _
+from django.utils.translation import gettext, gettext_lazy as _
 from django.utils import timezone
+from django.urls import reverse
 
 from django.forms.models import modelformset_factory
 
-from .compat import reverse
 from .models import Newsletter, Subscription, Submission
 from .forms import (
     SubscribeRequestForm, UserUpdateForm, UpdateRequestForm,
@@ -45,7 +45,7 @@ def is_authenticated(user):
     return user.is_authenticated if isinstance(user.is_authenticated, bool) else user.is_authenticated()
 
 
-class NewsletterViewBase(object):
+class NewsletterViewBase:
     """ Base class for newsletter views. """
     queryset = Newsletter.on_site.filter(visible=True)
     allow_empty = False
@@ -66,10 +66,10 @@ class NewsletterListView(NewsletterViewBase, ListView):
         """ Allow post requests. """
 
         # All logic (for now) occurs in the form logic
-        return super(NewsletterListView, self).get(request, **kwargs)
+        return super().get(request, **kwargs)
 
     def get_context_data(self, **kwargs):
-        context = super(NewsletterListView, self).get_context_data(**kwargs)
+        context = super().get_context_data(**kwargs)
 
         if is_authenticated(self.request.user):
             # Add a formset for logged in users.
@@ -115,7 +115,7 @@ class NewsletterListView(NewsletterViewBase, ListView):
 
                 messages.info(
                     request,
-                    ugettext("Your changes have been saved.")
+                    gettext("Your changes have been saved.")
                 )
 
             except ValidationError:
@@ -137,7 +137,7 @@ class NewsletterListView(NewsletterViewBase, ListView):
         return formset
 
 
-class ProcessUrlDataMixin(object):
+class ProcessUrlDataMixin:
     """
     Mixin providing the ability to process args and kwargs from url
     before dispatching request.
@@ -150,7 +150,7 @@ class ProcessUrlDataMixin(object):
     def dispatch(self, *args, **kwargs):
         self.process_url_data(*args, **kwargs)
 
-        return super(ProcessUrlDataMixin, self).dispatch(*args, **kwargs)
+        return super().dispatch(*args, **kwargs)
 
 
 class NewsletterMixin(ProcessUrlDataMixin):
@@ -167,7 +167,7 @@ class NewsletterMixin(ProcessUrlDataMixin):
 
         assert 'newsletter_slug' in kwargs
 
-        super(NewsletterMixin, self).process_url_data(*args, **kwargs)
+        super().process_url_data(*args, **kwargs)
 
         newsletter_queryset = kwargs.get(
             'newsletter_queryset',
@@ -181,7 +181,7 @@ class NewsletterMixin(ProcessUrlDataMixin):
 
     def get_form_kwargs(self):
         """ Add newsletter to form kwargs. """
-        kwargs = super(NewsletterMixin, self).get_form_kwargs()
+        kwargs = super().get_form_kwargs()
 
         kwargs['newsletter'] = self.newsletter
 
@@ -189,7 +189,7 @@ class NewsletterMixin(ProcessUrlDataMixin):
 
     def get_context_data(self, **kwargs):
         """ Add newsletter to context. """
-        context = super(NewsletterMixin, self).get_context_data(**kwargs)
+        context = super().get_context_data(**kwargs)
 
         context['newsletter'] = self.newsletter
 
@@ -203,7 +203,7 @@ class ActionMixin(ProcessUrlDataMixin):
 
     def process_url_data(self, *args, **kwargs):
         """ Add action from url to instance attributes if not already set. """
-        super(ActionMixin, self).process_url_data(*args, **kwargs)
+        super().process_url_data(*args, **kwargs)
 
         if self.action is None:
             assert 'action' in kwargs
@@ -213,7 +213,7 @@ class ActionMixin(ProcessUrlDataMixin):
 
     def get_context_data(self, **kwargs):
         """ Add action to context. """
-        context = super(ActionMixin, self).get_context_data(**kwargs)
+        context = super().get_context_data(**kwargs)
 
         context['action'] = self.action
 
@@ -276,7 +276,7 @@ class ActionUserView(ActionTemplateView):
 
     def process_url_data(self, *args, **kwargs):
         """ Add confirm to instance attributes. """
-        super(ActionUserView, self).process_url_data(*args, **kwargs)
+        super().process_url_data(*args, **kwargs)
 
         # confirm is optional kwarg defaulting to False
         self.confirm = kwargs.get('confirm', False)
@@ -286,7 +286,7 @@ class ActionUserView(ActionTemplateView):
 
     @method_decorator(login_required)
     def dispatch(self, *args, **kwargs):
-        return super(ActionUserView, self).dispatch(*args, **kwargs)
+        return super().dispatch(*args, **kwargs)
 
 
 class SubscribeUserView(ActionUserView):
@@ -323,7 +323,7 @@ class SubscribeUserView(ActionUserView):
                 _('You are already subscribed to %s.') % self.newsletter
             )
 
-        return super(SubscribeUserView, self).get(request, *args, **kwargs)
+        return super().get(request, *args, **kwargs)
 
 
 class UnsubscribeUserView(ActionUserView):
@@ -365,7 +365,7 @@ class UnsubscribeUserView(ActionUserView):
                 _('You are not subscribed to %s.') % self.newsletter
             )
 
-        return super(UnsubscribeUserView, self).get(request, *args, **kwargs)
+        return super().get(request, *args, **kwargs)
 
 
 class ActionRequestView(ActionFormView):
@@ -374,13 +374,13 @@ class ActionRequestView(ActionFormView):
 
     def process_url_data(self, *args, **kwargs):
         """ Add error to instance attributes. """
-        super(ActionRequestView, self).process_url_data(*args, **kwargs)
+        super().process_url_data(*args, **kwargs)
 
         self.error = None
 
     def get_context_data(self, **kwargs):
         """ Add error to context. """
-        context = super(ActionRequestView, self).get_context_data(**kwargs)
+        context = super().get_context_data(**kwargs)
 
         context.update({
             'error': self.error,
@@ -427,9 +427,9 @@ class ActionRequestView(ActionFormView):
 
             # Although form was valid there was error while sending email,
             # so stay at the same url.
-            return super(ActionRequestView, self).form_invalid(form)
+            return super().form_invalid(form)
 
-        return super(ActionRequestView, self).form_valid(form)
+        return super().form_valid(form)
 
 
 class SubscribeRequestView(ActionRequestView):
@@ -439,7 +439,7 @@ class SubscribeRequestView(ActionRequestView):
 
     def get_form_kwargs(self):
         """ Add ip to form kwargs for submitted forms. """
-        kwargs = super(SubscribeRequestView, self).get_form_kwargs()
+        kwargs = super().get_form_kwargs()
 
         if self.request.method in ('POST', 'PUT'):
             kwargs['ip'] = self.request.META.get('REMOTE_ADDR')
@@ -454,7 +454,7 @@ class SubscribeRequestView(ActionRequestView):
             kwargs['confirm'] = self.confirm
             return SubscribeUserView.as_view()(request, *args, **kwargs)
 
-        return super(SubscribeRequestView, self).dispatch(
+        return super().dispatch(
             request, *args, **kwargs
         )
 
@@ -469,7 +469,7 @@ class UnsubscribeRequestView(ActionRequestView):
             kwargs['confirm'] = self.confirm
             return UnsubscribeUserView.as_view()(request, *args, **kwargs)
 
-        return super(UnsubscribeRequestView, self).dispatch(
+        return super().dispatch(
             request, *args, **kwargs
         )
 
@@ -494,7 +494,7 @@ class UpdateSubscriptionView(ActionFormView):
         """
         assert 'email' in kwargs
 
-        super(UpdateSubscriptionView, self).process_url_data(*args, **kwargs)
+        super().process_url_data(*args, **kwargs)
 
         self.subscription = get_object_or_404(
             Subscription, newsletter=self.newsletter,
@@ -513,7 +513,7 @@ class UpdateSubscriptionView(ActionFormView):
 
     def get_form_kwargs(self):
         """ Add instance to form kwargs. """
-        kwargs = super(UpdateSubscriptionView, self).get_form_kwargs()
+        kwargs = super().get_form_kwargs()
 
         kwargs['instance'] = self.subscription
 
@@ -528,7 +528,7 @@ class UpdateSubscriptionView(ActionFormView):
 
         subscription.update(self.action)
 
-        return super(UpdateSubscriptionView, self).form_valid(form)
+        return super().form_valid(form)
 
 
 class SubmissionViewBase(NewsletterMixin):
@@ -547,12 +547,11 @@ class SubmissionViewBase(NewsletterMixin):
         """ Use only visible newsletters. """
 
         kwargs['newsletter_queryset'] = NewsletterListView().get_queryset()
-        return super(
-            SubmissionViewBase, self).process_url_data(*args, **kwargs)
+        return super().process_url_data(*args, **kwargs)
 
     def get_queryset(self):
         """ Filter out submissions for current newsletter. """
-        qs = super(SubmissionViewBase, self).get_queryset()
+        qs = super().get_queryset()
 
         qs = qs.filter(newsletter=self.newsletter)
 
@@ -566,7 +565,7 @@ class SubmissionViewBase(NewsletterMixin):
         default time zone, so that displayed items are consistent with the URL.
 
         Related discussion:
-        https://github.com/dokterbob/django-newsletter/issues/74
+        https://github.com/jazzband/django-newsletter/issues/74
         """
         value = datetime.datetime.combine(value, datetime.time.min)
         if settings.USE_TZ:
@@ -586,12 +585,23 @@ class SubmissionArchiveDetailView(SubmissionViewBase, DateDetailView):
         Make sure the actual message is available.
         """
         context = \
-            super(SubmissionArchiveDetailView, self).get_context_data(**kwargs)
+            super().get_context_data(**kwargs)
 
         message = self.object.message
 
+        # Determines the appropriate template to display a thumbnail
+        if newsletter_settings.THUMBNAIL == 'sorl-thumbnail':
+            thumbnail_template = (
+                'newsletter/message/thumbnail/sorl_thumbnail.html'
+            )
+        elif newsletter_settings.THUMBNAIL == 'easy-thumbnails':
+            thumbnail_template = (
+                'newsletter/message/thumbnail/easy_thumbnails.html'
+            )
+
         context.update({
             'message': message,
+            'attachment_links': True,
             'site': Site.objects.get_current(),
             'date': self.object.publish_date,
             'STATIC_URL': settings.STATIC_URL,
@@ -599,6 +609,7 @@ class SubmissionArchiveDetailView(SubmissionViewBase, DateDetailView):
             # custom template extends from base template which needs request
             # to be in the context
             'request': self.request,  # custom
+            'thumbnail_template': thumbnail_template,
         })
 
         return context
@@ -613,7 +624,7 @@ class SubmissionArchiveDetailView(SubmissionViewBase, DateDetailView):
 
         # No HTML -> no party!
         if not html_template:
-            raise Http404(ugettext(
+            raise Http404(gettext(
                 'No HTML template associated with the newsletter this '
                 'message belongs to.'
             ))