+++ tests/test_web.py
@@ -1,29 +1,21 @@
-from datetime import datetime, timedelta
-
 import time
+import pytz
 
 import unittest
+from unittest.mock import patch, PropertyMock
 
-# Conditionally import pytz
-try:
-    import pytz
-except ImportError:
-    pytz = None
-
-from django.contrib.auth import get_user_model
+from datetime import datetime, timedelta
 
 from django.core import mail
-
+from django.contrib.auth import get_user_model
 from django.utils import timezone
 from django.utils.encoding import force_text
-
 from django.test.utils import override_settings
+from django.urls import reverse
 
 from newsletter.models import (
     Newsletter, Subscription, Submission, Message, get_default_sites
 )
-
-from newsletter.compat import reverse
 from newsletter.forms import UpdateForm
 
 from .utils import MailTestCase, UserTestCase, WebTestCase, ComparingTestCase
@@ -346,7 +338,7 @@ class SubscribeTestCase(WebTestCase, MailTestCase):
                     kwargs={'newsletter_slug': self.n.slug,
                             'action': 'unsubscribe'})
 
-        super(SubscribeTestCase, self).setUp()
+        super().setUp()
 
     def test_urls(self):
         # TODO: is performing this test in each subclass
@@ -1258,7 +1250,7 @@ class InvisibleAnonymousSubscribeTestCase(AnonymousSubscribeTestCase):
     """
 
     def setUp(self):
-        super_obj = super(InvisibleAnonymousSubscribeTestCase, self)
+        super_obj = super()
         super_obj.setUp()
 
         # Make newsletter invisible
@@ -1276,7 +1268,7 @@ class InvisibleUserSubscribeTestCase(UserSubscribeTestCase):
     """
 
     def setUp(self):
-        super_obj = super(InvisibleUserSubscribeTestCase, self)
+        super_obj = super()
         super_obj.setUp()
 
         # Make newsletter invisible
@@ -1346,6 +1338,40 @@ class ArchiveTestcase(NewsletterListTestCase):
 
         self.assertContains(response, self.submission.message.title)
 
+    @patch(
+        'newsletter.settings.NewsletterSettings.THUMBNAIL',
+        new_callable=PropertyMock,
+    )
+    def test_archive_detail_sorl_thumbnail_template(self, THUMBNAIL):
+        """Tests that sorl-thumbnail template works."""
+        THUMBNAIL.return_value = 'sorl-thumbnail'
+
+        detail_url = self.submission.get_absolute_url()
+
+        response = self.client.get(detail_url)
+        self.assertEqual(response.status_code, 200)
+
+        self.assertTemplateUsed(
+            'newsletter/message/thumbnail/sorl_thumbnail.html'
+        )
+
+    @patch(
+        'newsletter.settings.NewsletterSettings.THUMBNAIL',
+        new_callable=PropertyMock,
+    )
+    def test_archive_detail_easy_thumbnails_template(self, THUMBNAIL):
+        """Tests that easy-thumbnails template works."""
+        THUMBNAIL.return_value = 'easy-thumbnails'
+
+        detail_url = self.submission.get_absolute_url()
+
+        response = self.client.get(detail_url)
+        self.assertEqual(response.status_code, 200)
+
+        self.assertTemplateUsed(
+            'newsletter/message/thumbnail/easy_thumbnails.html'
+        )
+
     def test_archive_unpublished_detail(self):
         """ Assert that an unpublished submission is truly inaccessible. """
 
@@ -1375,7 +1401,7 @@ class ArchiveTestcase(NewsletterListTestCase):
 
         Ref:
             * https://docs.djangoproject.com/en/1.5/topics/i18n/timezones/#troubleshooting
-            * https://github.com/dokterbob/django-newsletter/issues/74
+            * https://github.com/jazzband/django-newsletter/issues/74
         """
         problematic_date = datetime(2012, 3, 3, 1, 30)
 
@@ -1394,7 +1420,7 @@ class ArchiveTestcase(NewsletterListTestCase):
         self.test_archive_detail()
 
 
-class ActionTemplateViewMixin(object):
+class ActionTemplateViewMixin:
     """ Mixin for testing requests to urls for all three actions. """
 
     def get_action_url(self, action):