+++ CHANGES.rst
@@ -1,6 +1,20 @@
 Changes
 =======
 
+1.0b1 (unreleased)
+------------------
+- Add support for Django 3.1
+- Configurable thumbnailing, dropping hard sorl-thumbnail (#304).
+- File attachments for messages (#334).
+- Drop surlex dependency improved `path()` and `re_path()` (#339).
+- ...
+
+0.9.1 (18-05-2020)
+------------------
+
+- Drop support for Django 1.11 LTS and Python 3.4 and 2.7.
+- Updated translations.
+
 0.9 (19-02-2019)
 ----------------